app/requirements.txt
fastapi==0.75.0
pydantic==1.9.0
pymongo==4.0.2
uvicorn==0.17.6
dnspython==2.2.1
mangum==0.14.1
python-dotenv==0.20.0
mongoengine==0.24.1
pytest==7.1.2
mongomock==4.0.0
requests==2.27.1
